Working with neDB
Saving User Data and created a phone and saved it's data to the user
array of phones. Everything worked so far.

Still have to work through the actual GSX process and updating the UI
for amount of phones.

Maybe add a UI tab to view all phones that were worked on